requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
diffusers==0.6.0
    # via -r requirements.in
filelock==3.8.0
    # via
    #   diffusers
    #   huggingface-hub
    #   transformers
huggingface-hub==0.10.1
    # via
    #   diffusers
    #   transformers
idna==3.4
    # via requests
importlib-metadata==5.0.0
    # via diffusers
numpy==1.23.4
    # via
    #   diffusers
    #   transformers
packaging==21.3
    # via
    #   huggingface-hub
    #   transformers
pillow==9.2.0
    # via
    #   -r requirements.in
    #   diffusers
pyparsing==3.0.9
    # via packaging
pyyaml==6.0
    # via
    #   huggingface-hub
    #   transformers
regex==2022.9.13
    # via
    #   diffusers
    #   transformers
requests==2.28.1
    # via
    #   diffusers
    #   huggingface-hub
    #   transformers
tokenizers==0.13.1
    # via transformers
tqdm==4.64.1
    # via
    #   huggingface-hub
    #   transformers
transformers==4.23.1
    # via -r requirements.in
typing-extensions==4.4.0
    # via huggingface-hub
urllib3==1.26.12
    # via requests
zipp==3.9.0
    # via importlib-metadata
